requirements.txt
Cython==0.29.26
cython-bbox==0.1.3
lap==0.4.0
numpy==1.20.2
opencv-python==4.5.4.60
opencv_contrib_python==4.5.4.60
opencv-python-headless==4.5.4.60
PyYAML==6.0
scipy==1.7.2
torch==1.8.0
